Implement multi-file Obj-C interop tests (interop_objc_tests task)
Extract most of interop_objc_smoke tests
