+++ backend.native/tests/interop/objc/objcTests.def
@@ -1,0 +1,6 @@
+language = Objective-C
+headers = Foundation/NSArray.h Foundation/NSValue.h Foundation/NSString.h Foundation/NSStream.h Foundation/NSBundle.h
+headerFilter = **/interop/objc/tests/**.h \
+    Foundation/NSArray.h Foundation/NSValue.h Foundation/NSString.h Foundation/NSStream.h \
+    objc/objc.h Foundation/NSBundle.h
+linkerOpts = -lobjctests
