+++ backend.native/tests/interop/objc/smoke.h
@@ -76,24 +76,6 @@ id createObjectWithFactory(id<ObjectFactory> factory) {
   return [factory create];
 }
 
-@protocol BlockProvider
-@required
--(int (^)(int)) block;
-@end;
-
-int callProvidedBlock(id<BlockProvider> blockProvider, int argument) {
-  return [blockProvider block](argument);
-}
-
-@protocol BlockConsumer
-@required
--(int)callBlock:(int (^)(int))block argument:(int)argument;
-@end;
-
-int callPlusOneBlock(id<BlockConsumer> blockConsumer, int argument) {
-  return [blockConsumer callBlock:^int(int p) { return p + 1; } argument:argument];
-}
-
 @protocol CustomRetainMethods
 @required
 -(id)returnRetained:(id)obj __attribute__((ns_returns_retained));
@@ -126,86 +108,6 @@ void useCustomRetainMethods(id<CustomRetainMethods> p) {
   [p returnRetainedBlock:retainedBlock]();
 }
 
-NSObject* createNSObject() {
-  return [NSObject new];
-}
-
-@protocol ExceptionThrower
--(void)throwException;
-@end;
-
-@interface ExceptionThrowerManager : NSObject
-+(void)throwExceptionWith:(id<ExceptionThrower>)thrower;
-@end;
-
-@interface Blocks : NSObject
-+(BOOL)blockIsNull:(void (^)(void))block;
-+(int (^)(int, int, int, int))same:(int (^)(int, int, int, int))block;
-
-@property (class) void (^nullBlock)(void);
-@property (class) void (^notNullBlock)(void);
-@end;
-
-@interface TestVarargs : NSObject
--(instancetype _Nonnull)initWithFormat:(NSString*)format, ...;
-+(instancetype _Nonnull)testVarargsWithFormat:(NSString*)format, ...;
-@property NSString* formatted;
-
-+(NSString* _Nonnull)stringWithFormat:(NSString*)format, ...;
-+(NSObject* _Nonnull)stringWithFormat:(NSString*)format args:(void*)args;
-@end;
-
-@interface TestVarargs (TestVarargsExtension)
--(instancetype _Nonnull)initWithFormat:(NSString*)format, ...;
-@end;
-
-@interface TestVarargsSubclass : TestVarargs
-// Test clashes:
--(instancetype _Nonnull)initWithFormat:(NSString*)format args:(void*)args;
-+(NSString* _Nonnull)stringWithFormat:(NSString*)format args:(void*)args;
-@end;
-
-@interface TestOverrideInit : NSObject
--(instancetype)initWithValue:(int)value NS_DESIGNATED_INITIALIZER;
-+(instancetype)createWithValue:(int)value;
-@end;
-
-@interface MultipleInheritanceClashBase : NSObject
-@property (nonnull) MultipleInheritanceClashBase* delegate;
-@end;
-
-@protocol MultipleInheritanceClash
-@optional
-@property (nullable) id<MultipleInheritanceClash> delegate;
-@end;
-
-@interface MultipleInheritanceClash1 : MultipleInheritanceClashBase <MultipleInheritanceClash>
-@end;
-
-@interface MultipleInheritanceClash2 : MultipleInheritanceClashBase <MultipleInheritanceClash>
-@property MultipleInheritanceClashBase* delegate;
-@end;
-
-@interface TestClashingWithAny1 : NSObject
--(NSString*)toString;
--(NSString*)toString_;
--(int)hashCode;
--(BOOL)equals:(id _Nullable)other;
-@end;
-
-@interface TestClashingWithAny2 : NSObject
--(void)toString;
--(void)hashCode;
--(void)equals:(int)p; // May clash.
-@end;
-
-@interface TestClashingWithAny3 : NSObject
-// Not clashing actually.
--(NSString*)toString:(int)p;
--(int)hashCode:(int)p;
--(BOOL)equals;
-@end;
-
 id getPrinterProtocolRaw() {
   return @protocol(Printer);
 }
@@ -213,76 +115,3 @@ id getPrinterProtocolRaw() {
 Protocol* getPrinterProtocol() {
   return @protocol(Printer);
 }
-
-@interface TestInitWithCustomSelector : NSObject
--(instancetype)initCustom;
-@property BOOL custom;
-
-+(instancetype _Nonnull)createCustom;
-@end;
-
-@interface TestAllocNoRetain : NSObject
-@property BOOL ok;
-@end;
-
-@interface CustomString : NSString
-- (instancetype)initWithValue:(int)value;
-@property int value;
-@end;
-
-CustomString* _Nonnull createCustomString(int value) {
-    return [[CustomString alloc] initWithValue:value];
-}
-
-int getCustomStringValue(CustomString* str) {
-    return str.value;
-}
-
-extern BOOL customStringDeallocated;
-
-@interface TestConstructorReturnsNull : NSObject
-- (instancetype)init;
-@end;
-
-@interface TestCallableReferences : NSObject
-@property int value;
-- (int)instanceMethod;
-+ (int)classMethod:(int)first :(int)second;
-@end;
-
-// [KT-36067] cinterop tool fails when there is a structure member named Companion
-
-struct EnumFieldMangleStruct {
-	enum {Companion, Any} smth;
-};
-struct EnumFieldMangleStruct enumMangledStruct = { Any };
-
-struct MyStruct {
-	int Companion;      // simple clash
-	int _Companion;
-	int $_Companion;
-	int super;
-};
-
-struct MyStruct myStruct = {11, 12, 13, 14};
-
-@protocol Proto
-@property int Companion;  // clash on implementing
-@end;
-
-@interface FooMangled : NSObject<Proto>
-//- (void) CompanionS;  // mangleSimple does not support this: it may clash after mangling
-@end;
-
-
-@class DeallocListener;
-
-@interface DeallocExecutor : NSObject
-@property DeallocListener* deallocListener;
-@end;
-
-@interface DeallocListener : NSObject
-@property (weak) DeallocExecutor* deallocExecutor;
-@property BOOL deallocated;
--(BOOL)deallocExecutorIsNil;
-@end;
