+++ backend.native/tests/interop/objc/smoke.kt
@@ -16,27 +16,11 @@ fun main(args: Array<String>) {
 }
 
 fun run() {
-    testConversions()
+    // TODO: migrate remaining tests to interop/objc/tests/
     testTypeOps()
-    testWeakRefs()
-    testExceptions()
-    testBlocks()
     testCustomRetain()
-    testVarargs()
-    testOverrideInit()
-    testMultipleInheritanceClash()
-    testClashingWithAny()
-    testInitWithCustomSelector()
-    testAllocNoRetain()
-    testNSOutputStreamToMemoryConstructor()
     testExportObjCClass()
-    testCustomString()
     testLocalizedStrings()
-    testConstructorReturnsNull()
-    testCallableReferences()
-    testMangling()
-    testSharing()
-    testObjCWeakRef()
 
     assertEquals(2, ForwardDeclaredEnum.TWO.value)
 
@@ -197,80 +181,6 @@ fun testTypeOps() {
     assertFails { MutablePairImpl(1, 2).asAny() as Foo }
 }
 
-fun testConversions() {
-    testMethodsOfAny(emptyList<Nothing>(), NSArray())
-    testMethodsOfAny(listOf(1, "foo"), nsArrayOf(1, "foo"))
-    testMethodsOfAny(42, NSNumber.numberWithInt(42), 17)
-    testMethodsOfAny(true, NSNumber.numberWithBool(true), false)
-}
-
-fun testMethodsOfAny(kotlinObject: Any, equalNsObject: NSObject, otherObject: Any = Any()) {
-    assertEquals(kotlinObject.hashCode(), equalNsObject.hashCode())
-    assertEquals(kotlinObject.toString(), equalNsObject.toString())
-    assertEquals(kotlinObject, equalNsObject)
-    assertEquals(equalNsObject, kotlinObject)
-    assertNotEquals(equalNsObject, otherObject)
-}
-
-fun testWeakRefs() {
-    testWeakReference({ createNSObject()!! })
-
-    createAndAbandonWeakRef(NSObject())
-
-    testWeakReference({ NSArray.arrayWithArray(listOf(42)) as NSArray })
-}
-
-fun testWeakReference(block: () -> NSObject) {
-    val ref = autoreleasepool {
-        createAndTestWeakReference(block)
-    }
-
-    kotlin.native.internal.GC.collect()
-
-    assertNull(ref.get())
-}
-
-fun createAndTestWeakReference(block: () -> NSObject): WeakReference<NSObject> {
-    val ref = createWeakReference(block)
-    assertNotNull(ref.get())
-    assertEquals(ref.get()!!.hash(), ref.get()!!.hash())
-    return ref
-}
-
-fun createWeakReference(block: () -> NSObject) = WeakReference(block())
-
-fun createAndAbandonWeakRef(obj: NSObject) {
-    WeakReference(obj)
-}
-
-fun testExceptions() {
-    assertFailsWith<MyException> {
-        ExceptionThrowerManager.throwExceptionWith(object : NSObject(), ExceptionThrowerProtocol {
-            override fun throwException() {
-                throw MyException()
-            }
-        })
-    }
-}
-
-fun testBlocks() {
-    assertTrue(Blocks.blockIsNull(null))
-    assertFalse(Blocks.blockIsNull({}))
-
-    assertEquals(null, Blocks.nullBlock)
-    assertNotEquals(null, Blocks.notNullBlock)
-
-    assertEquals(10, Blocks.same({ a, b, c, d -> a + b + c + d })!!(1, 2, 3, 4))
-
-    assertEquals(222, callProvidedBlock(object : NSObject(), BlockProviderProtocol {
-        override fun block(): (Int) -> Int = { it * 2 }
-    }, 111))
-
-    assertEquals(322, callPlusOneBlock(object : NSObject(), BlockConsumerProtocol {
-        override fun callBlock(block: ((Int) -> Int)?, argument: Int) = block!!(argument)
-    }, 321))
-}
-
 private lateinit var retainedMustNotBeDeallocated: MustNotBeDeallocated
 
 fun testCustomRetain() {
@@ -299,156 +209,6 @@ fun testCustomRetain() {
     assertFalse(unexpectedDeallocation)
 }
 
-fun testVarargs() {
-    assertEquals(
-            "a b -1",
-            TestVarargs.testVarargsWithFormat(
-                    "%@ %s %d",
-                    "a" as NSString, "b".cstr, (-1).toByte()
-            ).formatted
-    )
-
-    assertEquals(
-            "2 3 9223372036854775807",
-            TestVarargs(
-                    "%d %d %lld",
-                    2.toShort(), 3, Long.MAX_VALUE
-            ).formatted
-    )
-
-    assertEquals(
-            "0.1 0.2 1 0",
-            TestVarargs.create(
-                    "%.1f %.1lf %d %d",
-                    0.1.toFloat(), 0.2, true, false
-            ).formatted
-    )
-
-    assertEquals(
-            "1 2 3",
-            TestVarargs(
-                    format = "%d %d %d",
-                    args = *arrayOf(1, 2, 3)
-            ).formatted
-    )
-
-    assertEquals(
-            "4 5 6",
-            TestVarargs(
-                    args = *arrayOf(4, *arrayOf(5, 6)),
-                    format = "%d %d %d"
-            ).formatted
-    )
-
-    assertEquals(
-            "7",
-            TestVarargsSubclass.stringWithFormat(
-                    "%d",
-                    7
-            )
-    )
-}
-
-fun testOverrideInit() {
-    assertEquals(42, (TestOverrideInitImpl.createWithValue(42) as TestOverrideInitImpl).value)
-}
-
-class TestOverrideInitImpl @OverrideInit constructor(val value: Int) : TestOverrideInit(value) {
-    companion object : TestOverrideInitMeta()
-}
-
-private class MyException : Throwable()
-
-fun testMultipleInheritanceClash() {
-    val clash1 = MultipleInheritanceClash1()
-    val clash2 = MultipleInheritanceClash2()
-
-    clash1.delegate = clash1
-    assertEquals(clash1, clash1.delegate)
-    clash1.setDelegate(clash2)
-    assertEquals(clash2, clash1.delegate())
-
-    clash2.delegate = clash1
-    assertEquals(clash1, clash2.delegate)
-    clash2.setDelegate(clash2)
-    assertEquals(clash2, clash2.delegate())
-}
-
-fun testClashingWithAny() {
-    assertEquals("description", TestClashingWithAny1().toString())
-    assertEquals("toString", TestClashingWithAny1().toString_())
-    assertEquals("toString_", TestClashingWithAny1().toString__())
-    assertEquals(1, TestClashingWithAny1().hashCode())
-    assertEquals(31, TestClashingWithAny1().hashCode_())
-    assertFalse(TestClashingWithAny1().equals(TestClashingWithAny1()))
-    assertTrue(TestClashingWithAny1().equals_(TestClashingWithAny1()))
-
-    assertEquals("description", TestClashingWithAny2().toString())
-    assertEquals(Unit, TestClashingWithAny2().toString_())
-    assertEquals(2, TestClashingWithAny2().hashCode())
-    assertEquals(Unit, TestClashingWithAny2().hashCode_())
-    assertFalse(TestClashingWithAny2().equals(TestClashingWithAny2()))
-    assertEquals(Unit, TestClashingWithAny2().equals_(42))
-
-    assertEquals("description", TestClashingWithAny3().toString())
-    assertEquals("toString:11", TestClashingWithAny3().toString(11))
-    assertEquals(3, TestClashingWithAny3().hashCode())
-    assertEquals(4, TestClashingWithAny3().hashCode(3))
-    assertFalse(TestClashingWithAny3().equals(TestClashingWithAny3()))
-    assertTrue(TestClashingWithAny3().equals())
-}
-
-fun testInitWithCustomSelector() {
-    assertFalse(TestInitWithCustomSelector().custom)
-    assertTrue(TestInitWithCustomSelector(custom = Unit).custom)
-
-    val customSubclass: TestInitWithCustomSelector = TestInitWithCustomSelectorSubclass.createCustom()
-    assertTrue(customSubclass is TestInitWithCustomSelectorSubclass)
-    assertTrue(customSubclass.custom)
-
-    // Test side effect:
-    var ok = false
-    assertTrue(TestInitWithCustomSelector(run { ok = true }).custom)
-    assertTrue(ok)
-}
-
-private class TestInitWithCustomSelectorSubclass : TestInitWithCustomSelector {
-    @OverrideInit constructor(custom: Unit) : super(custom) {
-        assertSame(Unit, custom)
-    }
-
-    companion object : TestInitWithCustomSelectorMeta()
-}
-
-fun testAllocNoRetain() {
-    // Ensure that calling Kotlin constructor generated for Objective-C initializer doesn't result in
-    // redundant retain-release sequence for `alloc` result, since it may provoke specific bugs to reproduce, e.g.
-    // the one found in [[NSOutputStream alloc] initToMemory] sequence where initToMemory deallocates its receiver
-    // forcibly when replacing it with other object: (to be compiled with ARC enabled)
-    /*
-    #import <Foundation/Foundation.h>
-
-    void* mem;
-    NSOutputStream* allocated = nil;
-
-    int main() {
-        allocated = [NSOutputStream alloc];
-        NSOutputStream* initialized = [allocated initToMemory];
-        mem = calloc(1, 0x10); // To corrupt the 'allocated' object header.
-        allocated = nil; // Crashes here in objc_release.
-
-        return 0;
-    }
-     */
-
-    assertTrue(TestAllocNoRetain().ok)
-}
-
-fun testNSOutputStreamToMemoryConstructor() {
-    val stream: Any = NSOutputStream(toMemory = Unit)
-    assertTrue(stream is NSOutputStream)
-}
-
 private const val TestExportObjCClass1Name = "TestExportObjCClass"
 @ExportObjCClass(TestExportObjCClass1Name) class TestExportObjCClass1 : NSObject()
 
@@ -466,21 +226,6 @@ fun testExportObjCClass() {
             xor (TestExportObjCClass4().objCClassName == TestExportObjCClass34Name))
 }
 
-fun testCustomString() {
-    assertFalse(customStringDeallocated)
-
-    fun test() = autoreleasepool {
-        val str: String = createCustomString(321)
-        assertEquals("321", str)
-        assertEquals("CustomString", str.objCClassName)
-        assertEquals(321, getCustomStringValue(str))
-    }
-
-    test()
-    kotlin.native.internal.GC.collect()
-    assertTrue(customStringDeallocated)
-}
-
 fun testLocalizedStrings() {
     val key = "screen_main_plural_string"
     val localizedString = NSBundle.mainBundle.localizedStringForKey(key, value = "", table = "Localizable")
@@ -488,140 +233,7 @@ fun testLocalizedStrings() {
     assertEquals("Plural: 5 apples", string)
 }
 
-fun testConstructorReturnsNull() {
-    assertFailsWith<NullPointerException>() {
-        TestConstructorReturnsNull()
-    }
-}
-
-fun testCallableReferences() {
-    val createTestCallableReferences = ::TestCallableReferences
-    assertEquals("<init>", createTestCallableReferences.name)
-    val testCallableReferences: Any = createTestCallableReferences()
-    assertTrue(testCallableReferences is TestCallableReferences)
-
-    val valueRef: kotlin.reflect.KMutableProperty0<Int> = testCallableReferences::value
-    assertEquals("value", valueRef.name)
-    assertEquals(0, valueRef())
-    valueRef.set(42)
-    assertEquals(42, valueRef())
-
-    val classMethodRef = (TestCallableReferences)::classMethod
-    assertEquals("classMethod", classMethodRef.name)
-    assertEquals(3, classMethodRef(1, 2))
-
-    val instanceMethodRef = TestCallableReferences::instanceMethod
-    assertEquals("instanceMethod", instanceMethodRef.name)
-    assertEquals(42, instanceMethodRef(testCallableReferences))
-
-    val boundInstanceMethodRef = testCallableReferences::instanceMethod
-    assertEquals("instanceMethod", boundInstanceMethodRef.name)
-    assertEquals(42, boundInstanceMethodRef())
-}
-
-fun testMangling() {
-    assertEquals(11, myStruct.`Companion$`)
-    assertEquals(12, myStruct._Companion)
-    assertEquals(13, myStruct.`$_Companion`)
-    assertEquals(14, myStruct.`super`)
-
-    val objc = FooMangled()
-    objc.`Companion$` = 99
-    assertEquals(99, objc.Companion())
-    assertEquals(99, objc.`Companion$`)
-
-    enumMangledStruct.smth = Companion
-    assertEquals(Companion, enumMangledStruct.smth)
-}
-
-private class NSObjectImpl : NSObject() {
-    var x = 111
-}
-
-fun testSharing() = withWorker {
-    val obj = NSObjectImpl()
-    val array = nsArrayOf(obj)
-
-    assertFalse(obj.isFrozen)
-
-    runInWorker {
-        assertFailsWith<IncorrectDereferenceException> {
-            array.objectAtIndex(0)
-        }
-    }
-
-    obj.x = 222
-    obj.freeze()
-    assertTrue(obj.isFrozen)
-
-    runInWorker {
-        val obj1 = array.objectAtIndex(0) as NSObjectImpl
-        assertFailsWith<InvalidMutabilityException> {
-            obj1.x = 333
-        }
-    }
-
-    assertEquals(222, obj.x)
-
-    // TODO: test [obj release] etc.
-}
-
-fun testObjCWeakRef() {
-    val deallocListener = DeallocListener()
-    assertFalse(deallocListener.deallocated)
-
-    testObjCWeakRef0(deallocListener)
-
-    kotlin.native.internal.GC.collect()
-    assertTrue(deallocListener.deallocated)
-    assertTrue(deallocListener.deallocExecutorIsNil())
-}
-
-fun testObjCWeakRef0(deallocListener: DeallocListener) = withWorker {
-    assertTrue(deallocListener.deallocExecutorIsNil())
-
-    val obj = object : DeallocExecutor() {}
-    deallocListener.deallocExecutor = obj
-    obj.deallocListener = deallocListener
-
-    assertFalse(deallocListener.deallocExecutorIsNil())
-
-//    TODO: can't actually test, Obj-C runtime doesn't expect _tryRetain throwing an exception.
-//    runInWorker {
-//        assertFailsWith<IncorrectDereferenceException> {
-//            deallocListener.deallocExecutorIsNil()
-//        }
-//    }
-
-    obj.freeze()
-
-    runInWorker {
-        // [deallocListener.deallocExecutorIsNil()] calls deallocExecutor getter, which retains [obj] and either
-        // puts it to autoreleasepool or releases it immediately (Obj-C ARC optimization).
-        // Wrap the call to autoreleasepool to ensure [obj] will be released:
-        autoreleasepool {
-            assertFalse(deallocListener.deallocExecutorIsNil())
-        }
-        // Process release of Kotlin reference to [obj] in any case:
-        kotlin.native.internal.GC.collect()
-    }
-}
-
-private fun Worker.runInWorker(block: () -> Unit) {
-    block.freeze()
-    val future = this.execute(TransferMode.SAFE, { block }) {
-        it()
-    }
-    future.result // Throws on failure.
-}
-
 private val Any.objCClassName: String
     get() = object_getClassName(this)!!.toKString()
-
-fun nsArrayOf(vararg elements: Any): NSArray = NSMutableArray().apply {
-    elements.forEach {
-        this.addObject(it as ObjCObject)
-    }
-}
 
 fun Any?.asAny(): Any? = this
