+++ backend.native/tests/interop/objc/tests/allocNoRetain.h
@@ -1,0 +1,5 @@
+#import <Foundation/NSObject.h>
+
+@interface TestAllocNoRetain : NSObject
+@property BOOL ok;
+@end;
