+++ backend.native/tests/interop/objc/tests/callableReferences.h
@@ -1,0 +1,7 @@
+#import <Foundation/NSObject.h>
+
+@interface TestCallableReferences : NSObject
+@property int value;
+- (int)instanceMethod;
++ (int)classMethod:(int)first :(int)second;
+@end;
