+++ backend.native/tests/interop/objc/tests/callableReferences.m
@@ -1,0 +1,11 @@
+#import "callableReferences.h"
+
+@implementation TestCallableReferences
+- (int)instanceMethod {
+    return self.value;
+}
+
++ (int)classMethod:(int)first :(int)second {
+    return first + second;
+}
+@end;
