+++ backend.native/tests/interop/objc/tests/constructorReturnsNull.h
@@ -1,0 +1,5 @@
+#import <Foundation/NSObject.h>
+
+@interface TestConstructorReturnsNull : NSObject
+- (instancetype)init;
+@end;
