+++ backend.native/tests/interop/objc/tests/constructorReturnsNull.kt
@@ -1,0 +1,9 @@
+import kotlinx.cinterop.*
+import kotlin.test.*
+import objcTests.*
+
+@Test fun testConstructorReturnsNull() {
+    assertFailsWith<NullPointerException>() {
+        TestConstructorReturnsNull()
+    }
+}
