+++ backend.native/tests/interop/objc/tests/constructorReturnsNull.m
@@ -1,0 +1,7 @@
+#import "constructorReturnsNull.h"
+
+@implementation TestConstructorReturnsNull
+- (instancetype)init {
+    return nil;
+}
+@end;
