+++ backend.native/tests/interop/objc/tests/customString.h
@@ -1,0 +1,16 @@
+#import <Foundation/NSString.h>
+
+@interface CustomString : NSString
+- (instancetype)initWithValue:(int)value;
+@property int value;
+@end;
+
+CustomString* _Nonnull createCustomString(int value) {
+    return [[CustomString alloc] initWithValue:value];
+}
+
+int getCustomStringValue(CustomString* str) {
+    return str.value;
+}
+
+extern BOOL customStringDeallocated;
