+++ backend.native/tests/interop/objc/tests/exceptions.h
@@ -1,0 +1,9 @@
+#import <Foundation/NSObject.h>
+
+@protocol ExceptionThrower
+-(void)throwException;
+@end;
+
+@interface ExceptionThrowerManager : NSObject
++(void)throwExceptionWith:(id<ExceptionThrower>)thrower;
+@end;
