+++ backend.native/tests/interop/objc/tests/exceptions.m
@@ -1,0 +1,7 @@
+#import "exceptions.h"
+
+@implementation ExceptionThrowerManager
++(void)throwExceptionWith:(id<ExceptionThrower>)thrower {
+    [thrower throwException];
+}
+@end;
