+++ backend.native/tests/interop/objc/tests/initWithCustomSelector.h
@@ -1,0 +1,8 @@
+#import <Foundation/NSObject.h>
+
+@interface TestInitWithCustomSelector : NSObject
+-(instancetype)initCustom;
+@property BOOL custom;
+
++(instancetype _Nonnull)createCustom;
+@end;
