+++ backend.native/tests/interop/objc/tests/main.kt
@@ -1,0 +1,11 @@
+import kotlin.system.exitProcess
+import kotlinx.cinterop.autoreleasepool
+import kotlin.native.internal.test.testLauncherEntryPoint
+
+fun main(args: Array<String>) {
+    autoreleasepool {
+        val exitCode = testLauncherEntryPoint(args)
+        // Note: this test runner checks for memory leaks after successful execution, unlike standard one.
+        if (exitCode != 0) exitProcess(exitCode)
+    }
+}
