+++ backend.native/tests/interop/objc/tests/mangling.m
@@ -1,0 +1,6 @@
+#import "mangling.h"
+
+// [KT-36067] mangling
+@implementation FooMangled : NSObject
+@synthesize Companion;
+@end;
