+++ backend.native/tests/interop/objc/tests/multipleInheritanceClash.m
@@ -1,0 +1,10 @@
+#import "multipleInheritanceClash.h"
+
+@implementation MultipleInheritanceClashBase
+@end;
+
+@implementation MultipleInheritanceClash1
+@end;
+
+@implementation MultipleInheritanceClash2
+@end;
