+++ backend.native/tests/interop/objc/tests/objcWeakRefs.h
@@ -1,0 +1,13 @@
+#import <Foundation/NSObject.h>
+
+@class DeallocListener;
+
+@interface DeallocExecutor : NSObject
+@property DeallocListener* deallocListener;
+@end;
+
+@interface DeallocListener : NSObject
+@property (weak) DeallocExecutor* deallocExecutor;
+@property BOOL deallocated;
+-(BOOL)deallocExecutorIsNil;
+@end;
