+++ backend.native/tests/interop/objc/tests/objcWeakRefs.m
@@ -1,0 +1,13 @@
+#import "objcWeakRefs.h"
+
+@implementation DeallocExecutor
+-(void)dealloc {
+    self.deallocListener.deallocated = YES;
+}
+@end;
+
+@implementation DeallocListener
+-(BOOL)deallocExecutorIsNil {
+    return self.deallocExecutor == nil;
+}
+@end;
