+++ backend.native/tests/interop/objc/tests/overrideInit.h
@@ -1,0 +1,6 @@
+#import <Foundation/NSObject.h>
+
+@interface TestOverrideInit : NSObject
+-(instancetype)initWithValue:(int)value NS_DESIGNATED_INITIALIZER;
++(instancetype)createWithValue:(int)value;
+@end;
