+++ backend.native/tests/interop/objc/tests/overrideInit.m
@@ -1,0 +1,11 @@
+#import "OverrideInit.h"
+
+@implementation TestOverrideInit
+-(instancetype)initWithValue:(int)value {
+    return self = [super init];
+}
+
++(instancetype)createWithValue:(int)value {
+    return [[self alloc] initWithValue:value];
+}
+@end;
