+++ backend.native/tests/interop/objc/tests/sharing.kt
@@ -1,0 +1,36 @@
+import kotlinx.cinterop.*
+import kotlin.native.concurrent.*
+import kotlin.test.*
+import objcTests.*
+
+private class NSObjectImpl : NSObject() {
+    var x = 111
+}
+
+@Test fun testSharing() = withWorker {
+    val obj = NSObjectImpl()
+    val array = nsArrayOf(obj)
+
+    assertFalse(obj.isFrozen)
+
+    runInWorker {
+        assertFailsWith<IncorrectDereferenceException> {
+            array.objectAtIndex(0)
+        }
+    }
+
+    obj.x = 222
+    obj.freeze()
+    assertTrue(obj.isFrozen)
+
+    runInWorker {
+        val obj1 = array.objectAtIndex(0) as NSObjectImpl
+        assertFailsWith<InvalidMutabilityException> {
+            obj1.x = 333
+        }
+    }
+
+    assertEquals(222, obj.x)
+
+    // TODO: test [obj release] etc.
+}
