+++ backend.native/tests/interop/objc/tests/weakRefs.h
@@ -1,0 +1,5 @@
+#import <Foundation/NSObject.h>
+
+NSObject* createNSObject() {
+  return [NSObject new];
+}
